Drop unneeded `RECIPE_ROOT`
The `RECIPE_ROOT` variable was not set after moving the tests into the
package name. However this is not needed anyways as we already instruct
`conda-build` to copy the test file into the test directory. So it
should already be in the current working directory. Given this, just
drop it.